# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Symbiota API",
        "contact": {
            "name": "Symbiota Support Hub Team",
            "email": "symbiota@asu.edu"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "portal"
        }
    ],
    "paths": {
        "/api/v2/installation": {
            "get": {
                "tags": [
                    ""
                ],
                "operationId": "/api/v2/installation",
                "responses": {
                    "200": {
                        "description": "Returns list of installations registered within system",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. "
                    }
                }
            }
        },
        "/api/v2/installation/{identifier}": {
            "get": {
                "tags": [
                    ""
                ],
                "operationId": "/api/v2/installation/identifier",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "Installation ID or GUID associated with target installation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns metabase on installation registered within system with matching ID",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. Installation identifier is required."
                    }
                }
            }
        },
        "/api/v2/installation/ping": {
            "get": {
                "tags": [
                    ""
                ],
                "operationId": "/api/v2/installation/ping",
                "responses": {
                    "200": {
                        "description": "Returns installation metadata",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. "
                    }
                }
            }
        },
        "/api/v2/installation/{identifier}/touch": {
            "get": {
                "tags": [
                    ""
                ],
                "operationId": "/api/v2/installation/identifier/touch",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "Identifier of the remote installation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "urlPath",
                        "in": "query",
                        "description": "Url to Symbiota root of remote installation",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns metabase remote installation, if successfully registered",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. Identifier of remote installation is required."
                    }
                }
            }
        },
        "/api/v2/occurrence": {
            "get": {
                "tags": [
                    ""
                ],
                "operationId": "/api/v2/occurrence",
                "parameters": [
                    {
                        "name": "catalogNumber",
                        "in": "query",
                        "description": "catalogNumber",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "occurrenceID",
                        "in": "query",
                        "description": "occurrenceID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "country",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "stateProvince",
                        "in": "query",
                        "description": "State, Province, or second level political unit",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "county",
                        "in": "query",
                        "description": "County, parish, or third level political unit",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns list of occurrences",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. "
                    }
                }
            }
        },
        "/api/v2/occurrence/{identifier}": {
            "get": {
                "tags": [
                    ""
                ],
                "operationId": "/api/v2/occurrence/identifier",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "occid or specimen GUID (occurrenceID) associated with target occurrence",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "includeMedia",
                        "in": "query",
                        "description": "Whether to include media within output",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns occurrence data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. Occurrence identifier is required."
                    }
                }
            }
        }
    }
}